requirements.txt
jira
requests
python-dotenv
